Move likelihood.cpp completely to the header.
